# openapi.yaml
openapi: 3.0.0
info:
  title: Feedback 360 Survey API
  version: 2.0.0
  description: |
    This API specification contains information around the Feedback 360 Survey
    API developed by the App Dev Practice.

    The Feedback 360 Survey API enables the management of feedback surveys for
    customer engagements. A feedback survey is N/A to 5 point scale of competency around
    the technologies used for a project. Each added employee will be able to enhance the
    survey by adding technologies, softskills, and freetext for each of his or her
    colleagues. The will enable Red Hat to better understand multiple aspects of
    a consultant's skillset.
  contact:
    email: appdevpractice@redhat.com
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
...
servers:
  - url: "http://localhost:3007"
    description: LOCAL
  - url: "http://survey-service-consultant-360-dev.apps.shared-dev.dev.openshift.opentlc.com"
    description: Spring Boot Service that implements the Feedback 360 Survey API on DEV Cluster
  - url: "http://survey-service-consultant-360-prod.apps.shared-dev.dev.openshift.opentlc.com"
    description: Spring Boot Service that implements the Feedback 360 Survey API on PROD Cluster
paths:
...
  /userprofile:
    get:
      operationId: getUserProfile
      description: For the currently authenticated user, return the user profile information
      tags:
        - userProfile
      responses:
        "200":
          description: The available user information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Employee"
      tags:
        - userProfile
      responses:
        "200":
          description: >-
            The user profile has been updated with the contents of this request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "400":
          description: >-
            The user profile was not updated. Something is wrong with the request like a missing
            required field.
        "404":
          description: This user profile was not found
      operationId: updateUserProfile
      summary: |
        Update the user profile resource
      description: >
        The representation of the user profile contained in this request will replace the user
        profile. This operation is idempotent.
...
components:
  schemas:
...
    Employee:
      type: object
      x-java-class-annotations:
        - "@javax.persistence.MappedSuperclass"
      x-tablename: employees
      properties:
        id:
          description: Employee's RedHat Kerberos ID
          type: string
          example: jsmith@redhat.com
          x-java-field-annotations:
            - "@javax.persistence.Id"
          x-primary-key: true
        name:
          description: Name of the employee.
          type: string
          example: John Smith
        email:
          description: The email address of the employee.
          type: string
          example: jsmith@redhat.com
          x-java-field-annotation:
            - |-
              "@org.hibernate.annotations.Formula("CONCAT(id, '@redhat.com')")"
        role:
          description: "The role of the employee. For example, consultant, PM, TSM, etc."
          type: string
          example: Consultant
        hasAcceptedTerms:
          description: Whether the employee has accepted the terms of the usage agreement.
          type: boolean
          example: false
          x-java-field-annotations:
            - "@javax.persistence.Transient"
        dateTermsAccepted:
          description: The date the employee accepted the terms of the usage agreement.
          type: string
          format: date
          example: "2021-02-09"
        termsVersionNumber:
          description: The version number of terms of the usage agreement.
          type: number
          format: float
          example: 1.0
...